Move core service to backend
