+++ backend/package.json
@@ -16,7 +16,18 @@
     "@activitypods/marketplace": "1.1.0",
     "@activitypods/messages": "1.1.0",
     "@activitypods/profiles": "1.1.0",
+    "@activitypods/synchronizer": "1.1.0",
+    "@semapps/activitypub": "0.6.0-alpha.0",
+    "@semapps/auth": "0.6.0-alpha.0",
+    "@semapps/jsonld": "0.6.0-alpha.0",
+    "@semapps/ldp": "0.6.0-alpha.0",
     "@semapps/notifications": "0.6.0-alpha.0",
+    "@semapps/pod": "0.6.0-alpha.0",
+    "@semapps/signature": "0.6.0-alpha.0",
+    "@semapps/sparql-endpoint": "0.6.0-alpha.0",
+    "@semapps/triplestore": "0.6.0-alpha.0",
+    "@semapps/webfinger": "0.6.0-alpha.0",
+    "@semapps/webid": "0.6.0-alpha.0",
     "@semapps/sync": "0.6.0-alpha.0",
     "@semapps/webacl": "0.6.0-alpha.0",
     "dotenv-flow": "^3.2.0",
