+++ backend/services/synchronizer.service.js
@@ -1,0 +1,11 @@
+const { SynchronizerService } = require('@activitypods/synchronizer');
+
+module.exports = {
+  mixins: [SynchronizerService],
+  settings: {
+    podProvider: true,
+    mirrorGraph: false,
+    synchronizeContainers: false,
+    attachToLocalContainers: true
+  }
+};
