--- packages/core/README.md
@@ -1,18 +1,0 @@
-# Core Services
-
-Core [ActivityPods](../../README.md) services
-
-## Services
-
-- [ActivityPubService](https://semapps.org/docs/middleware/activitypub/index)
-- [TripleStoreService](https://semapps.org/docs/middleware/triplestore)
-- [LdpService](https://semapps.org/docs/middleware/ldp/index)
-- [WebAclService](https://semapps.org/docs/middleware/webacl)
-- [WebfingerService](https://semapps.org/docs/middleware/webfinger)
-- [WebIdService](https://semapps.org/docs/middleware/webid)
-- [AuthLocalService](https://semapps.org/docs/middleware/auth)
-- [SignatureService](https://semapps.org/docs/middleware/signature)
-- SparqlEndpointService
-- PodService
-- JsonLdService
-- ProxyService
