--- packages/core/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  CoreService: require('./service')
-};
