--- packages/core/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@activitypods/core",
-  "version": "1.1.0",
-  "description": "ActivityPods core services",
-  "license": "Apache-2.0",
-  "author": "Virtual Assembly",
-  "dependencies": {
-    "@activitypods/announcer": "1.1.0",
-    "@semapps/activitypub": "0.6.0-alpha.0",
-    "@semapps/auth": "0.6.0-alpha.0",
-    "@semapps/jsonld": "0.6.0-alpha.0",
-    "@semapps/ldp": "0.6.0-alpha.0",
-    "@semapps/pod": "0.6.0-alpha.0",
-    "@semapps/signature": "0.6.0-alpha.0",
-    "@semapps/sparql-endpoint": "0.6.0-alpha.0",
-    "@semapps/sync": "0.6.0-alpha.0",
-    "@semapps/triplestore": "0.6.0-alpha.0",
-    "@semapps/webacl": "0.6.0-alpha.0",
-    "@semapps/webfinger": "0.6.0-alpha.0",
-    "@semapps/webid": "0.6.0-alpha.0",
-    "moleculer": "^0.14.19",
-    "moleculer-web": "^0.10.0-beta1",
-    "url-join": "^4.0.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
